+++ src/Structures/CustomBoardTrigger.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Graphics;
 using QuestFramework.Framework.Menus;
 using StardewModdingAPI;
 using StardewValley.Menus;
@@ -6,13 +7,14 @@ using System;
 
 namespace QuestFramework.Framework.Structures
 {
-    public class CustomBoardTrigger
+    public sealed class CustomBoardTrigger
     {
         public string LocationName { get; set; }
         public Point Tile { get; set; }
         public string BoardName { get; set; }
         public bool ShowIndicator { get; set; } = true;
         public BoardType BoardType { get; set; } = BoardType.Quests;
+        public Texture2D Texture { get; set; }
 
         public Func<bool> unlockConditionFunc;